requirements.txt
flask==2.3.2
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-exporter-otlp==1.18.0